+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - PlayWright tests
 - Allow manually entering bounding boxes for search
+- Support negating CQL2 filters (globally and per filter)
+- Support CQL2 Advanced Comparison Operators
+- Support CQL2 Array Functions
 
 ### Changed
 
@@ -41,6 +44,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Handle state of downloads better and confirm leaving the page when downloading
 - Better error on request to the /collections or .../items endpoints
 - Collection list on Global Item Search was empty in certain situations
+- Show an error message when no operator is supported for a queryable
 - Don't show an "unsupported" error when only Collection Search is supported by the API
 
 ## [4.0.1] - 2026-02-11